big code refactor: cocksize and pidor game - ended
